Test changes based on review feedback
